Revert "Merge branch 'master' into patch-1"
This reverts commit e1e9844b4088a205f66b1da25400dca55db9dd2b, reversing
changes made to 26952e35b441440d5ec823ca3eec0f5f94c90e68.

+++ src/libslic3r/SLA/SLARaster.hpp
@@ -68,14 +68,15 @@ public:
 
     /// Type that represents a resolution in pixels.
     struct Resolution {
-        size_t width_px;
-        size_t height_px;
+        unsigned width_px;
+        unsigned height_px;
 
-        inline Resolution(size_t w = 0, size_t h = 0)
-            : width_px(w), height_px(h)
-        {}
+        inline Resolution(unsigned w = 0, unsigned h = 0):
+            width_px(w), height_px(h) {}
 
-        inline size_t pixels() const { return width_px * height_px; }
+        inline unsigned pixels() const /*noexcept*/ {
+            return width_px * height_px;
+        }
     };
 
     /// Types that represents the dimension of a pixel in millimeters.
